Refactor line break regex into a more human readable form, adds/fixes a bit of cases
