+++ packages/excalidraw/element/textElement.ts
@@ -16,12 +16,12 @@ import {
   BOUND_TEXT_PADDING,
   DEFAULT_FONT_FAMILY,
   DEFAULT_FONT_SIZE,
-  ENV,
   TEXT_ALIGN,
   VERTICAL_ALIGN,
 } from "../constants";
 import type { MaybeTransformHandleType } from "./transformHandles";
 import { isTextElement } from ".";
+import { wrapText } from "./textWrapping";
 import { isBoundToContainer, isArrowElement } from "./typeChecks";
 import { LinearElementEditor } from "./linearElementEditor";
 import type { AppState } from "../types";
@@ -30,162 +30,6 @@ import {
   updateOriginalContainerCache,
 } from "./containerCache";
 import type { ExtractSetType } from "../utility-types";
-
-/**
- * Matches various emoji types.
- *
- * 1. basic emojis (😀, 🌍)
- * 2. flags (🇨🇿)
- * 3. multi-codepoint emojis:
- *    - skin tones (👍🏽)
- *    - variation selectors (☂️)
- *    - keycaps (1️⃣)
- *    - tag sequences (🏴󠁧󠁢󠁥󠁮󠁧󠁿)
- *    - emoji sequences (👨‍👩‍👧‍👦, 👩‍🚀, 🏳️‍🌈)
- *
- * Unicode points:
- * - \uFE0F: presentation selector
- * - \u20E3: enclosing keycap
- * - \u200D: ZWJ (zero width joiner)
- * - \u{E0020}-\u{E007E}: tags
- * - \u{E007F}: cancel tag
- *
- * @see https://unicode.org/reports/tr51/#EBNF_and_Regex, with changes:
- * - replaced \p{Emoji} with [\p{Extended_Pictographic}\p{Emoji_Presentation}], see more in `should tokenize emojis mixed with mixed text` test
- * - replaced \p{Emod} with \p{Emoji_Modifier} as some do not understand the abbreviation (i.e. https://devina.io/redos-checker)
- */
-const _EMOJI_CHAR =
-  /(\p{RI}\p{RI}|[\p{Extended_Pictographic}\p{Emoji_Presentation}](?:\p{Emoji_Modifier}|\uFE0F\u20E3?|[\u{E0020}-\u{E007E}]+\u{E007F})?(?:\u200D(?:\p{RI}\p{RI}|[\p{Emoji}](?:\p{Emoji_Modifier}|\uFE0F\u20E3?|[\u{E0020}-\u{E007E}]+\u{E007F})?))*)/u;
-
-/**
- * Detect a CJK char, though does not include every possible char used in CJK texts,
- * such as symbols and punctuations.
- *
- * By default every CJK is a breaking point, though CJK has additional breaking points,
- * including full width punctuations or symbols (Chinese and Japanese) and western punctuations (Korean).
- *
- * Additional CJK breaking point rules:
- * - expect a break before (lookahead), but not after (negative lookbehind),  i.e. "(" or "("
- * - expect a break after (lookbehind), but not before (negative lookahead), i.e. "）" or ")"
- * - expect a break always (lookahead and lookbehind), i.e. "〃"
- */
-const _CJK_CHAR =
-  /\p{Script=Han}\p{Script=Hiragana}\p{Script=Katakana}\p{Script=Hangul}/u;
-const _CJK_SYMBOLS_AND_PUNCTUATION =
-  /（）［］｛｝〈〉《》｟｠｢｣「」『』【】〖〗〔〕〘〙〚〛＜＞〝〞＇〟・。ﾟﾞ，、．：；？！％ー/u;
-/**
- * Following characters break with any character, even though are mostly used with CJK.
- *
- * Hello た。→ ["Hello", "た。"]
- *        ↑ DON'T BREAK "た。" (negative lookahead)
- * Hello「た」 World → ["Hello", "「た」", "World"]
- *       ↑ DON'T BREAK "「た" (negative lookbehind)
- *        ↑ DON'T BREAK "た」"(negative lookahead)
- *      ↑ BREAK BEFORE "「" (lookahead)
- *         ↑ BREAK AFTER "」" (lookbehind)
- */
-const _CJK_BREAK_NOT_AFTER_BUT_BEFORE =
-  /（［｛〈《｟｢「『【〖〔〘〚＜〝<\(\[\{/u;
-const _CJK_BREAK_NOT_BEFORE_BUT_AFTER =
-  /）］｝〉》｠｣」』】〗〕〙〛＞〞＇〟・。ﾟﾞ，、．：；？！％±‥…\/>\)\]\}.,:;\?!/u;
-const _CJK_BREAK_ALWAYS = /　〃〜～〰＃＆＊＋－ー／＝｜￢￣￤/u;
-
-/**
- * Natural breaking points for any grammars.
- *
- * Hello-world
- *       ↑ BREAK AFTER "-" → ["Hello-", "world"]
- * Hello world
- *      ↑ BREAK ALWAYS " " → ["Hello", " ", "world"]
- */
-const _ANY_BREAK_AFTER = /-/u;
-const _ANY_BREAK_ALWAYS = /\s/u;
-
-/**
- * Simple fallback for browsers (mainly Safari < 16.4) that don't support "Lookbehind assertion".
- *
- * Browser support as of 10/2024:
- * - 91% Lookbehind assertion https://caniuse.com/mdn-javascript_regular_expressions_lookbehind_assertion
- * - 94% Unicode character class escape https://caniuse.com/mdn-javascript_regular_expressions_unicode_character_class_escape
- *
- * Does not include advanced CJK breaking rules, but covers most of the core cases, especially for latin.
- */
-const BREAK_LINE_REGEX_SIMPLE = new RegExp(
-  `${_EMOJI_CHAR.source}|([${_ANY_BREAK_ALWAYS.source}${_CJK_CHAR.source}${_CJK_BREAK_ALWAYS.source}${_ANY_BREAK_AFTER.source}])`,
-  "u",
-);
-
-// Hello World → ["Hello", " World"]
-//      ↑ BREAK BEFORE " "
-// HelloたWorld → ["Hello", "たWorld"]
-//      ↑ BREAK BEFORE "た"
-// Hello「World」→ ["Hello", "「World」"]
-//      ↑ BREAK BEFORE "「"
-const getLookaheadBreakingPoints = () => {
-  const ANY_BREAKING_POINT = `(?=[${_ANY_BREAK_ALWAYS.source}])`;
-  const CJK_BREAKING_POINT = `(?<![${_CJK_BREAK_NOT_AFTER_BUT_BEFORE.source}])(?=[${_CJK_BREAK_NOT_AFTER_BUT_BEFORE.source}]*[${_CJK_CHAR.source}${_CJK_BREAK_ALWAYS.source}])`;
-  return new RegExp(`(?:${ANY_BREAKING_POINT}|${CJK_BREAKING_POINT})`, "u");
-};
-
-// Hello World → ["Hello ", "World"]
-//       ↑ BREAK AFTER " "
-// Hello-World → ["Hello-", "World"]
-//       ↑ BREAK AFTER "-"
-// HelloたWorld → ["Helloた", "World"]
-//       ↑ BREAK AFTER "た"
-//「Hello」World → ["「Hello」", "World"]
-//       ↑ BREAK AFTER "」"
-const getLookbehindBreakingPoints = () => {
-  const ANY_BREAKING_POINT = `(?<=[${_ANY_BREAK_ALWAYS.source}${_ANY_BREAK_AFTER.source}])`;
-  const CJK_BREAKING_POINT = `(?![${_CJK_BREAK_NOT_BEFORE_BUT_AFTER.source}${_ANY_BREAK_AFTER.source}])(?<=[${_CJK_CHAR.source}${_CJK_BREAK_ALWAYS.source}][${_CJK_BREAK_NOT_BEFORE_BUT_AFTER.source}]*)`;
-  return new RegExp(`(?:${ANY_BREAKING_POINT}|${CJK_BREAKING_POINT})`, "u");
-};
-
-/**
- * Break a line based on the whitespaces, CJK / emoji chars and language specific breaking points,
- * like hyphen for alphabetic and various full-width codepoints for CJK - especially Japanese, e.g.:
- *
- *  "Hello 世界。🌎🗺" → ["Hello", " ", "世", "界。", "🌎", "🗺"]
- *  "Hello-world" → ["Hello-", "world"]
- *  "「Hello World」" → ["「Hello", " ", "World」"]
- */
-const getBreakLineRegexAdvanced = () =>
-  new RegExp(
-    `${_EMOJI_CHAR.source}|${getLookaheadBreakingPoints().source}|${
-      getLookbehindBreakingPoints().source
-    }`,
-    "u",
-  );
-
-let cachedBreakLineRegex: RegExp | undefined;
-
-// Lazy-load for browsers that don't support "Lookbehind assertion"
-const getBreakLineRegex = () => {
-  if (!cachedBreakLineRegex) {
-    try {
-      cachedBreakLineRegex = getBreakLineRegexAdvanced();
-    } catch {
-      cachedBreakLineRegex = BREAK_LINE_REGEX_SIMPLE;
-    }
-  }
-
-  return cachedBreakLineRegex;
-};
-
-const CJK_REGEX = new RegExp(
-  `[${_CJK_CHAR.source}${_CJK_BREAK_ALWAYS.source}${_CJK_SYMBOLS_AND_PUNCTUATION.source}]`,
-  "u",
-);
-
-const EMOJI_REGEX = new RegExp(`${_EMOJI_CHAR.source}`, "u");
-
-export const containsCJK = (text: string) => {
-  return CJK_REGEX.test(text);
-};
-
-export const containsEmoji = (text: string) => {
-  return EMOJI_REGEX.test(text);
-};
 
 export const normalizeText = (text: string) => {
   return (
@@ -500,7 +344,7 @@ let canvas: HTMLCanvasElement | undefined;
  *
  * `Math.ceil` of the final width adds additional buffer which stabilizes slight wrapping incosistencies.
  */
-const getLineWidth = (
+export const getLineWidth = (
   text: string,
   font: FontString,
   forceAdvanceWidth?: true,
@@ -563,197 +407,6 @@ export const getTextHeight = (
 ) => {
   const lineCount = splitIntoLines(text).length;
   return getLineHeightInPx(fontSize, lineHeight) * lineCount;
-};
-
-export const parseTokens = (line: string) => {
-  const breakLineRegex = getBreakLineRegex();
-
-  // normalizing to single-codepoint composed chars due to canonical equivalence of multi-codepoint versions for chars like č, で (~ so that we don't break a line in between c and ˇ)
-  // filtering due to multi-codepoint chars like 👨‍👩‍👧‍👦, 👩🏽‍🦰
-  return line.normalize("NFC").split(breakLineRegex).filter(Boolean);
-};
-
-// handles multi-byte chars (é, 中) and purposefully does not handle multi-codepoint char (👨‍👩‍👧‍👦, 👩🏽‍🦰)
-const isSingleCharacter = (maybeSingleCharacter: string) => {
-  return (
-    maybeSingleCharacter.codePointAt(0) !== undefined &&
-    maybeSingleCharacter.codePointAt(1) === undefined
-  );
-};
-
-const satisfiesWordInvariant = (word: string) => {
-  if (import.meta.env.MODE === ENV.TEST || import.meta.env.DEV) {
-    if (/\s/.test(word)) {
-      throw new Error("Word should not contain any whitespaces!");
-    }
-  }
-};
-
-const wrapWord = (
-  word: string,
-  font: FontString,
-  maxWidth: number,
-): Array<string> => {
-  // multi-codepoint emojis are already broken apart and shouldn't be broken further
-  if (EMOJI_REGEX.test(word)) {
-    return [word];
-  }
-
-  satisfiesWordInvariant(word);
-
-  const lines: Array<string> = [];
-  const chars = Array.from(word);
-
-  let currentLine = "";
-  let currentLineWidth = 0;
-
-  for (const char of chars) {
-    const _charWidth = charWidth.calculate(char, font);
-    const testLineWidth = currentLineWidth + _charWidth;
-
-    if (testLineWidth <= maxWidth) {
-      currentLine = currentLine + char;
-      currentLineWidth = testLineWidth;
-      continue;
-    }
-
-    if (currentLine) {
-      lines.push(currentLine);
-    }
-
-    currentLine = char;
-    currentLineWidth = _charWidth;
-  }
-
-  if (currentLine) {
-    lines.push(currentLine);
-  }
-
-  return lines;
-};
-
-const wrapLine = (
-  line: string,
-  font: FontString,
-  maxWidth: number,
-): string[] => {
-  const lines: Array<string> = [];
-  const tokens = parseTokens(line);
-  const tokenIterator = tokens[Symbol.iterator]();
-
-  let currentLine = "";
-  let currentLineWidth = 0;
-
-  let iterator = tokenIterator.next();
-
-  while (!iterator.done) {
-    const token = iterator.value;
-    const testLine = currentLine + token;
-
-    // cache single codepoint whitespace, CJK or emoji width calc. as kerning should not apply here
-    const testLineWidth = isSingleCharacter(token)
-      ? currentLineWidth + charWidth.calculate(token, font)
-      : getLineWidth(testLine, font, true);
-
-    // build up the current line, skipping length check for possibly trailing whitespaces
-    if (/\s/.test(token) || testLineWidth <= maxWidth) {
-      currentLine = testLine;
-      currentLineWidth = testLineWidth;
-      iterator = tokenIterator.next();
-      continue;
-    }
-
-    // current line is empty => just the token (word) is longer than `maxWidth` and needs to be wrapped
-    if (!currentLine) {
-      const wrappedWord = wrapWord(token, font, maxWidth);
-      const trailingLine = wrappedWord[wrappedWord.length - 1] ?? "";
-      const precedingLines = wrappedWord.slice(0, -1);
-
-      lines.push(...precedingLines);
-
-      // trailing line of the wrapped word might -still be joined with next token/s
-      currentLine = trailingLine;
-      currentLineWidth = getLineWidth(trailingLine, font, true);
-      iterator = tokenIterator.next();
-    } else {
-      // push & reset, but don't iterate on the next token, as we didn't use it yet!
-      lines.push(currentLine.trimEnd());
-
-      // purposefully not iterating and not setting `currentLine` to `token`, so that we could use a simple !currentLine check above
-      currentLine = "";
-      currentLineWidth = 0;
-    }
-  }
-
-  // iterator done, push the trailing line if exists
-  if (currentLine) {
-    const trailingLine = trimTrailingLine(currentLine, font, maxWidth);
-    lines.push(trailingLine);
-  }
-
-  return lines;
-};
-
-// similarly to browsers, does not trim all whitespaces, but only those exceeding the maxWidth
-const trimTrailingLine = (line: string, font: FontString, maxWidth: number) => {
-  const shouldTrimWhitespaces = getLineWidth(line, font, true) > maxWidth;
-
-  if (!shouldTrimWhitespaces) {
-    return line;
-  }
-
-  // defensively default to `trimeEnd` in case the regex does not match
-  let [, trimmedLine, whitespaces] = line.match(/^(.+?)(\s+)$/) ?? [
-    line,
-    line.trimEnd(),
-    "",
-  ];
-
-  let trimmedLineWidth = getLineWidth(trimmedLine, font, true);
-
-  for (const whitespace of Array.from(whitespaces)) {
-    const _charWidth = charWidth.calculate(whitespace, font);
-    const testLineWidth = trimmedLineWidth + _charWidth;
-
-    if (testLineWidth > maxWidth) {
-      break;
-    }
-
-    trimmedLine = trimmedLine + whitespace;
-    trimmedLineWidth = testLineWidth;
-  }
-
-  return trimmedLine;
-};
-
-export const wrapText = (
-  text: string,
-  font: FontString,
-  maxWidth: number,
-): string => {
-  // if maxWidth is not finite or NaN which can happen in case of bugs in
-  // computation, we need to make sure we don't continue as we'll end up
-  // in an infinite loop
-  if (!Number.isFinite(maxWidth) || maxWidth < 0) {
-    return text;
-  }
-
-  const lines: Array<string> = [];
-  const originalLines = text.split("\n");
-
-  for (const originalLine of originalLines) {
-    const currentLineWidth = getLineWidth(originalLine, font, true);
-
-    if (currentLineWidth <= maxWidth) {
-      lines.push(originalLine);
-      continue;
-    }
-
-    const wrappedLine = wrapLine(originalLine, font, maxWidth);
-    lines.push(...wrappedLine);
-  }
-
-  return lines.join("\n");
 };
 
 export const charWidth = (() => {
